+++ x-pack/plugins/apm/public/components/app/Main/route_config/index.tsx
@@ -13,8 +13,8 @@ import { APMRouteDefinition } from '../../../../application/routes';
 import { toQuery } from '../../../shared/Links/url_helpers';
 import { ErrorGroupDetails } from '../../ErrorGroupDetails';
 import { Home } from '../../Home';
-import { ServiceDetails } from '../../ServiceDetails';
-import { ServiceNodeMetrics } from '../../ServiceNodeMetrics';
+import { ServiceDetails } from '../../service_details';
+import { ServiceNodeMetrics } from '../../service_node_metrics';
 import { Settings } from '../../Settings';
 import { AgentConfigurations } from '../../Settings/AgentConfigurations';
 import { AnomalyDetection } from '../../Settings/anomaly_detection';
