+++ x-pack/plugins/apm/public/components/shared/charts/helper/get_empty_series.ts
