+++ x-pack/plugins/apm/public/components/shared/charts/helper/get_timezone_offset_in_ms.ts
