+++ x-pack/plugins/apm/public/components/shared/charts/spark_plot/index.tsx
@@ -4,7 +4,13 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 import React from 'react';
-import { ScaleType, Chart, Settings, AreaSeries } from '@elastic/charts';
+import {
+  ScaleType,
+  Chart,
+  Settings,
+  AreaSeries,
+  CurveType,
+} from '@elastic/charts';
 import { EuiIcon } from '@elastic/eui';
 import { EuiFlexItem } from '@elastic/eui';
 import { EuiFlexGroup } from '@elastic/eui';
@@ -59,6 +65,7 @@ export function SparkPlot(props: Props) {
         yAccessors={['y']}
         data={series}
         color={color}
+        curve={CurveType.CURVE_MONOTONE_X}
       />
     </Chart>
   );
