+++ x-pack/plugins/apm/public/components/shared/charts/transaction_charts/index.tsx
@@ -26,10 +26,10 @@ import { ChartsSyncContextProvider } from '../../../../context/charts_sync_conte
 import { LicenseContext } from '../../../../context/LicenseContext';
 import { IUrlParams } from '../../../../context/UrlParamsContext/types';
 import { FETCH_STATUS } from '../../../../hooks/useFetcher';
-import { ITransactionChartData } from '../../../../selectors/chartSelectors';
+import { ITransactionChartData } from '../../../../selectors/chart_selectors';
 import { isValidCoordinateValue } from '../../../../utils/isValidCoordinateValue';
 import { TransactionBreakdown } from '../../TransactionBreakdown';
-import { LineChart } from '../line_chart';
+import { TimeseriesChart } from '../timeseries_chart';
 import { TransactionErrorRateChart } from '../transaction_error_rate_chart/';
 import { getResponseTimeTickFormatter } from './helper';
 import { MLHeader } from './ml_header';
@@ -81,7 +81,7 @@ export function TransactionCharts({
                   )}
                 </LicenseContext.Consumer>
               </EuiFlexGroup>
-              <LineChart
+              <TimeseriesChart
                 fetchStatus={fetchStatus}
                 id="transactionDuration"
                 timeseries={responseTimeSeries || []}
@@ -100,7 +100,7 @@ export function TransactionCharts({
               <EuiTitle size="xs">
                 <span>{tpmLabel(transactionType)}</span>
               </EuiTitle>
-              <LineChart
+              <TimeseriesChart
                 fetchStatus={fetchStatus}
                 id="requestPerMinutes"
                 timeseries={tpmSeries || []}
