+++ x-pack/plugins/apm/public/components/shared/charts/transaction_charts/use_formatter.ts
