+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -4942,7 +4942,6 @@
     "xpack.apm.metadataTable.section.urlLabel": "URL",
     "xpack.apm.metadataTable.section.userAgentLabel": "ユーザーエージェント",
     "xpack.apm.metadataTable.section.userLabel": "ユーザー",
-    "xpack.apm.metrics.plot.noDataLabel": "この時間範囲のデータがありません。",
     "xpack.apm.metrics.transactionChart.machineLearningLabel": "機械学習:",
     "xpack.apm.metrics.transactionChart.machineLearningTooltip": "平均期間の周りのストリームには予測バウンドが表示されます。異常スコアが>= 75の場合、注釈が表示されます。",
     "xpack.apm.metrics.transactionChart.machineLearningTooltip.withKuery": "フィルタリングで検索バーを使用しているときには、機械学習結果が表示されません",
@@ -5060,7 +5059,6 @@
     "xpack.apm.servicesTable.transactionsPerMinuteColumnLabel": "1 分あたりのトランザクション",
     "xpack.apm.servicesTable.transactionsPerMinuteUnitLabel": "1分あたりトランザクション数",
     "xpack.apm.servicesTable.UpgradeAssistantLink": "Kibana アップグレードアシスタントで詳細をご覧ください",
-    "xpack.apm.serviceVersion": "サービスバージョン",
     "xpack.apm.settings.agentConfig": "エージェントの編集",
     "xpack.apm.settings.anomaly_detection.legacy_jobs.body": "以前の統合のレガシー機械学習ジョブが見つかりました。これは、APMアプリでは使用されていません。",
     "xpack.apm.settings.anomaly_detection.legacy_jobs.button": "ジョブの確認",
@@ -5266,7 +5264,6 @@
     "xpack.apm.ux.percentile.label": "パーセンタイル",
     "xpack.apm.ux.title": "ユーザーエクスペリエンス",
     "xpack.apm.ux.visitorBreakdown.noData": "データがありません。",
-    "xpack.apm.version": "バージョン",
     "xpack.apm.waterfall.exceedsMax": "このトレースの項目数は表示されている範囲を超えています",
     "xpack.beatsManagement.beat.actionSectionTypeLabel": "タイプ: {beatType}。",
     "xpack.beatsManagement.beat.actionSectionVersionLabel": "バージョン: {beatVersion}",
