+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -4945,7 +4945,6 @@
     "xpack.apm.metadataTable.section.urlLabel": "URL",
     "xpack.apm.metadataTable.section.userAgentLabel": "用户代理",
     "xpack.apm.metadataTable.section.userLabel": "用户",
-    "xpack.apm.metrics.plot.noDataLabel": "此时间范围内没有数据。",
     "xpack.apm.metrics.transactionChart.machineLearningLabel": "Machine Learning",
     "xpack.apm.metrics.transactionChart.machineLearningTooltip": "环绕平均持续时间的流显示预期边界。对 ≥ 75 的异常分数显示标注。",
     "xpack.apm.metrics.transactionChart.machineLearningTooltip.withKuery": "使用搜索栏筛选时，Machine Learning 结果处于隐藏状态",
@@ -5064,7 +5063,6 @@
     "xpack.apm.servicesTable.transactionsPerMinuteColumnLabel": "每分钟事务数",
     "xpack.apm.servicesTable.transactionsPerMinuteUnitLabel": "tpm",
     "xpack.apm.servicesTable.UpgradeAssistantLink": "通过访问 Kibana 升级助手来了解详情",
-    "xpack.apm.serviceVersion": "服务版本",
     "xpack.apm.settings.agentConfig": "代理配置",
     "xpack.apm.settings.anomaly_detection.legacy_jobs.body": "我们在以前的集成中发现 APM 应用中不再使用的旧版 Machine Learning 作业",
     "xpack.apm.settings.anomaly_detection.legacy_jobs.button": "复查作业",
@@ -5271,7 +5269,6 @@
     "xpack.apm.ux.title": "用户体验",
     "xpack.apm.ux.url.hitEnter.include": "单击 {icon} 可包括与 {searchValue} 匹配的所有 URL",
     "xpack.apm.ux.visitorBreakdown.noData": "无数据。",
-    "xpack.apm.version": "版本",
     "xpack.apm.waterfall.exceedsMax": "此跟踪中的项目数超过显示的项目数",
     "xpack.beatsManagement.beat.actionSectionTypeLabel": "类型：{beatType}。",
     "xpack.beatsManagement.beat.actionSectionVersionLabel": "版本：{beatVersion}。",
